refactor: make content typesafe
