+++ src/_content/index.ts
@@ -1,4 +1,4 @@
-export * from './studies';
-export * from './testimonials';
-export * from './faqs';
-export * from './prices';
+export * from "./studies";
+export * from "./testimonials";
+export * from "./faqs";
+export * from "./prices";
