+++ src/app/businesses/page.tsx
@@ -3,7 +3,7 @@ import Hero from "~/components/Hero";
 import { Button, OutlineButton } from "~/components/Button";
 import BusinessSectionRow from "~/components/BusinessSectionRow";
 
-import caseStudies from "~/_content/studies.json";
+import { studies as caseStudies } from "~/_content";
 import CTA from "~/components/CTA";
 import CaseStudiesSection from "~/components/CaseStudiesSection";
