+++ src/app/case-studies/CaseGrid.tsx
@@ -1,7 +1,8 @@
 "use client";
 import { useState } from "react";
 import Container from "~/components/Container";
-import caseStudies from "~/_content/studies.json";
+// import caseStudies from "~/_content/studies.json";
+import { studies as caseStudies } from "~/_content";
 import { CaseStudyCardVariant } from "~/components/CaseStudyCard";
 
 export default function CaseGrid() {
